% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_spl_mask_gr}
\alias{EA_spl_mask_gr}
\title{Splice regions of all the transcripts in this package's transcript set, reduced to non-overlapping ranges.}
\description{
Splice regions of all the transcripts in this package's transcript set, reduced to non-overlapping ranges.
}
\details{
GRanges object containing splice regions. Transcripts where splice regions are overlapping are in the metadata.
Splice regions are 3bp of the exon and 8bp of the intron.
}
\author{
Dennis Kostka
}
\keyword{data}
